requirements.txt
APScheduler==3.5.3
obs-websocket-py==0.4
python-vlc==3.0.4106
pytz==2018.9
six==1.12.0
tzlocal==1.5.1
websocket-client==0.54.0
